Fix compiling errors
Add required _DISABLE_CONSTEXPR_MUTEX_CONSTRUCTOR to fix crash on load
Move EBlueprintType to resolve compiler error
Use bit_cast instead of C-Style case on Get/Set
Minor clean-up of UE.cpp
